migrated transformers dto model
