+++ internal/provider/job_resource.go
@@ -16,6 +16,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	mgmtv1alpha1 "github.com/nucleuscloud/neosync/backend/gen/go/protos/mgmt/v1alpha1"
 	"github.com/nucleuscloud/neosync/backend/gen/go/protos/mgmt/v1alpha1/mgmtv1alpha1connect"
+	transformer_model "github.com/nucleuscloud/terraform-provider-neosync/internal/models/transformers"
 )
 
 var _ resource.Resource = &JobResource{}
@@ -305,7 +306,7 @@ func fromModelJobMappings(input []*JobMapping) ([]*mgmtv1alpha1.JobMapping, erro
 				Config: &mgmtv1alpha1.TransformerConfig{},
 			},
 		}
-		config, err := fromModelTransformerConfig(inputMapping.Transformer.Config)
+		config, err := inputMapping.Transformer.ToDto()
 		if err != nil {
 			return nil, err
 		}
@@ -453,17 +454,16 @@ func fromJobDto(dto *mgmtv1alpha1.Job) (*JobResourceModel, error) {
 		model.Destinations = append(model.Destinations, dest)
 	}
 	for _, dtoMapping := range dto.Mappings {
-		tconfig, err := toTransformerConfigFromDto(dtoMapping.Transformer.Config)
+		tconfig := &transformer_model.Transformer{}
+		err := tconfig.FromDto(dtoMapping.Transformer.Config)
 		if err != nil {
 			return nil, err
 		}
 		mapping := &JobMapping{
-			Schema: types.StringValue(dtoMapping.Schema),
-			Table:  types.StringValue(dtoMapping.Table),
-			Column: types.StringValue(dtoMapping.Column),
-			Transformer: &Transformer{
-				Config: tconfig,
-			},
+			Schema:      types.StringValue(dtoMapping.Schema),
+			Table:       types.StringValue(dtoMapping.Table),
+			Column:      types.StringValue(dtoMapping.Column),
+			Transformer: tconfig,
 		}
 		model.Mappings = append(model.Mappings, mapping)
 	}
@@ -485,385 +485,6 @@ func fromJobDto(dto *mgmtv1alpha1.Job) (*JobResourceModel, error) {
 		}
 	}
 	return model, nil
-}
-
-func fromModelTransformerConfig(model *TransformerConfig) (*mgmtv1alpha1.TransformerConfig, error) {
-	dto := &mgmtv1alpha1.TransformerConfig{}
-	if model.GenerateEmail != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateEmailConfig{
-			GenerateEmailConfig: &mgmtv1alpha1.GenerateEmail{},
-		}
-	} else if model.TransformEmail != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_TransformEmailConfig{
-			TransformEmailConfig: &mgmtv1alpha1.TransformEmail{
-				PreserveDomain: model.TransformEmail.PreserveDomain.ValueBoolPointer(),
-				PreserveLength: model.TransformEmail.PreserveLength.ValueBoolPointer(),
-			},
-		}
-	} else if model.GenerateBool != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateBoolConfig{
-			GenerateBoolConfig: &mgmtv1alpha1.GenerateBool{},
-		}
-	} else if model.GenerateCardNumber != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateCardNumberConfig{
-			GenerateCardNumberConfig: &mgmtv1alpha1.GenerateCardNumber{
-				ValidLuhn: model.GenerateCardNumber.ValidLuhn.ValueBoolPointer(),
-			},
-		}
-	} else if model.GenerateCity != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateCityConfig{
-			GenerateCityConfig: &mgmtv1alpha1.GenerateCity{},
-		}
-	} else if model.GenerateE164PhoneNumber != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateE164PhoneNumberConfig{
-			GenerateE164PhoneNumberConfig: &mgmtv1alpha1.GenerateE164PhoneNumber{
-				Min: model.GenerateE164PhoneNumber.Min.ValueInt64Pointer(),
-				Max: model.GenerateE164PhoneNumber.Max.ValueInt64Pointer(),
-			},
-		}
-	} else if model.GenerateFirstName != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateFirstNameConfig{
-			GenerateFirstNameConfig: &mgmtv1alpha1.GenerateFirstName{},
-		}
-	} else if model.GenerateFloat64 != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateFloat64Config{
-			GenerateFloat64Config: &mgmtv1alpha1.GenerateFloat64{
-				RandomizeSign: model.GenerateFloat64.RandomizeSign.ValueBoolPointer(),
-				Min:           model.GenerateFloat64.Min.ValueFloat64Pointer(),
-				Max:           model.GenerateFloat64.Max.ValueFloat64Pointer(),
-			},
-		}
-	} else if model.GenerateFullAddress != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateFullAddressConfig{
-			GenerateFullAddressConfig: &mgmtv1alpha1.GenerateFullAddress{},
-		}
-	} else if model.GenerateFullName != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateFullNameConfig{
-			GenerateFullNameConfig: &mgmtv1alpha1.GenerateFullName{},
-		}
-	} else if model.GenerateGender != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateGenderConfig{
-			GenerateGenderConfig: &mgmtv1alpha1.GenerateGender{
-				Abbreviate: model.GenerateGender.Abbreviate.ValueBoolPointer(),
-			},
-		}
-	} else if model.GenerateInt64PhoneNumber != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateInt64PhoneNumberConfig{
-			GenerateInt64PhoneNumberConfig: &mgmtv1alpha1.GenerateInt64PhoneNumber{},
-		}
-	} else if model.GenerateInt64 != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateInt64Config{
-			GenerateInt64Config: &mgmtv1alpha1.GenerateInt64{
-				RandomizeSign: model.GenerateInt64.RandomizeSign.ValueBoolPointer(),
-				Min:           model.GenerateInt64.Min.ValueInt64Pointer(),
-				Max:           model.GenerateInt64.Max.ValueInt64Pointer(),
-			},
-		}
-	} else if model.GenerateLastName != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateLastNameConfig{
-			GenerateLastNameConfig: &mgmtv1alpha1.GenerateLastName{},
-		}
-	} else if model.GenerateSha256Hash != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateSha256HashConfig{
-			GenerateSha256HashConfig: &mgmtv1alpha1.GenerateSha256Hash{},
-		}
-	} else if model.GenerateSsn != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateSsnConfig{
-			GenerateSsnConfig: &mgmtv1alpha1.GenerateSSN{},
-		}
-	} else if model.GenerateState != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateStateConfig{
-			GenerateStateConfig: &mgmtv1alpha1.GenerateState{},
-		}
-	} else if model.GenerateStreetAddress != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateStreetAddressConfig{
-			GenerateStreetAddressConfig: &mgmtv1alpha1.GenerateStreetAddress{},
-		}
-	} else if model.GenerateStringPhoneNumber != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateStringPhoneNumberConfig{
-			GenerateStringPhoneNumberConfig: &mgmtv1alpha1.GenerateStringPhoneNumber{
-				Min: model.GenerateStringPhoneNumber.Min.ValueInt64Pointer(),
-				Max: model.GenerateStringPhoneNumber.Max.ValueInt64Pointer(),
-			},
-		}
-	} else if model.GenerateString != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateStringConfig{
-			GenerateStringConfig: &mgmtv1alpha1.GenerateString{
-				Min: model.GenerateString.Min.ValueInt64Pointer(),
-				Max: model.GenerateString.Max.ValueInt64Pointer(),
-			},
-		}
-	} else if model.GenerateUnixtimestamp != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateUnixtimestampConfig{
-			GenerateUnixtimestampConfig: &mgmtv1alpha1.GenerateUnixTimestamp{},
-		}
-	} else if model.GenerateUsername != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateUsernameConfig{
-			GenerateUsernameConfig: &mgmtv1alpha1.GenerateUsername{},
-		}
-	} else if model.GenerateUtctimestamp != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateUtctimestampConfig{
-			GenerateUtctimestampConfig: &mgmtv1alpha1.GenerateUtcTimestamp{},
-		}
-	} else if model.GenerateUuid != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateUuidConfig{
-			GenerateUuidConfig: &mgmtv1alpha1.GenerateUuid{
-				IncludeHyphens: model.GenerateUuid.IncludeHyphens.ValueBoolPointer(),
-			},
-		}
-	} else if model.GenerateZipcode != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateZipcodeConfig{
-			GenerateZipcodeConfig: &mgmtv1alpha1.GenerateZipcode{},
-		}
-	} else if model.TransformE164PhoneNumber != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateE164PhoneNumberConfig{
-			GenerateE164PhoneNumberConfig: &mgmtv1alpha1.GenerateE164PhoneNumber{
-				Min: model.GenerateE164PhoneNumber.Min.ValueInt64Pointer(),
-				Max: model.GenerateE164PhoneNumber.Max.ValueInt64Pointer(),
-			},
-		}
-	} else if model.TransformFirstName != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateFirstNameConfig{
-			GenerateFirstNameConfig: &mgmtv1alpha1.GenerateFirstName{},
-		}
-	} else if model.TransformFloat64 != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_TransformFloat64Config{
-			TransformFloat64Config: &mgmtv1alpha1.TransformFloat64{
-				RandomizationRangeMin: model.TransformFloat64.RandomizationRangeMin.ValueFloat64Pointer(),
-				RandomizationRangeMax: model.TransformFloat64.RandomizationRangeMax.ValueFloat64Pointer(),
-			},
-		}
-	} else if model.TransformFullName != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_TransformFullNameConfig{
-			TransformFullNameConfig: &mgmtv1alpha1.TransformFullName{
-				PreserveLength: model.TransformFullName.PreserveLength.ValueBoolPointer(),
-			},
-		}
-	} else if model.TransformInt64PhoneNumber != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_TransformInt64PhoneNumberConfig{
-			TransformInt64PhoneNumberConfig: &mgmtv1alpha1.TransformInt64PhoneNumber{
-				PreserveLength: model.TransformInt64PhoneNumber.PreserveLength.ValueBoolPointer(),
-			},
-		}
-	} else if model.TransformInt64 != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateInt64Config{
-			GenerateInt64Config: &mgmtv1alpha1.GenerateInt64{
-				RandomizeSign: model.GenerateInt64.RandomizeSign.ValueBoolPointer(),
-				Min:           model.GenerateInt64.Min.ValueInt64Pointer(),
-				Max:           model.GenerateInt64.Max.ValueInt64Pointer(),
-			},
-		}
-	} else if model.TransformLastName != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_TransformLastNameConfig{
-			TransformLastNameConfig: &mgmtv1alpha1.TransformLastName{
-				PreserveLength: model.TransformLastName.PreserveLength.ValueBoolPointer(),
-			},
-		}
-	} else if model.TransformPhoneNumber != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_TransformPhoneNumberConfig{
-			TransformPhoneNumberConfig: &mgmtv1alpha1.TransformPhoneNumber{
-				PreserveLength: model.TransformPhoneNumber.PreserveLength.ValueBoolPointer(),
-			},
-		}
-	} else if model.TransformString != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_TransformStringConfig{
-			TransformStringConfig: &mgmtv1alpha1.TransformString{
-				PreserveLength: model.TransformString.PreserveLength.ValueBoolPointer(),
-			},
-		}
-	} else if model.Passthrough != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_PassthroughConfig{
-			PassthroughConfig: &mgmtv1alpha1.Passthrough{},
-		}
-	} else if model.Null != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_Nullconfig{
-			Nullconfig: &mgmtv1alpha1.Null{},
-		}
-	} else if model.UserDefinedTransformer != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_UserDefinedTransformerConfig{
-			UserDefinedTransformerConfig: &mgmtv1alpha1.UserDefinedTransformerConfig{
-				Id: model.UserDefinedTransformer.Id.ValueString(),
-			},
-		}
-	} else if model.GenerateDefault != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateDefaultConfig{
-			GenerateDefaultConfig: &mgmtv1alpha1.GenerateDefault{},
-		}
-	} else if model.TransformJavascript != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_TransformJavascriptConfig{
-			TransformJavascriptConfig: &mgmtv1alpha1.TransformJavascript{
-				Code: model.TransformJavascript.Code.ValueString(),
-			},
-		}
-	} else if model.GenerateCategorical != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateCategoricalConfig{
-			GenerateCategoricalConfig: &mgmtv1alpha1.GenerateCategorical{
-				Categories: model.GenerateCategorical.Categories.ValueStringPointer(),
-			},
-		}
-	} else if model.TransformCharacterScramble != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_TransformCharacterScrambleConfig{
-			TransformCharacterScrambleConfig: &mgmtv1alpha1.TransformCharacterScramble{},
-		}
-	} else if model.GenerateJavascript != nil {
-		dto.Config = &mgmtv1alpha1.TransformerConfig_GenerateJavascriptConfig{
-			GenerateJavascriptConfig: &mgmtv1alpha1.GenerateJavascript{
-				Code: model.GenerateJavascript.Code.ValueString(),
-			},
-		}
-	} else {
-		return nil, fmt.Errorf("transformer config is not currently supported by this provider: %w", errors.ErrUnsupported)
-	}
-
-	return dto, nil
-}
-
-func toTransformerConfigFromDto(dto *mgmtv1alpha1.TransformerConfig) (*TransformerConfig, error) {
-	tconfig := &TransformerConfig{}
-	switch config := dto.Config.(type) {
-	case *mgmtv1alpha1.TransformerConfig_GenerateEmailConfig:
-		tconfig.GenerateEmail = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_TransformEmailConfig:
-		tconfig.TransformEmail = &TransformEmail{
-			PreserveDomain: types.BoolPointerValue(config.TransformEmailConfig.PreserveDomain),
-			PreserveLength: types.BoolPointerValue(config.TransformEmailConfig.PreserveLength),
-		}
-	case *mgmtv1alpha1.TransformerConfig_GenerateBoolConfig:
-		tconfig.GenerateBool = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateCardNumberConfig:
-		tconfig.GenerateCardNumber = &GenerateCardNumber{
-			ValidLuhn: types.BoolPointerValue(config.GenerateCardNumberConfig.ValidLuhn),
-		}
-	case *mgmtv1alpha1.TransformerConfig_GenerateCityConfig:
-		tconfig.GenerateCity = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateE164PhoneNumberConfig:
-		tconfig.GenerateE164PhoneNumber = &GenerateE164PhoneNumber{
-			Min: types.Int64PointerValue(config.GenerateE164PhoneNumberConfig.Min),
-			Max: types.Int64PointerValue(config.GenerateE164PhoneNumberConfig.Max),
-		}
-	case *mgmtv1alpha1.TransformerConfig_GenerateFirstNameConfig:
-		tconfig.GenerateFirstName = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateFloat64Config:
-		tconfig.GenerateFloat64 = &GenerateFloat64{
-			RandomizeSign: types.BoolPointerValue(config.GenerateFloat64Config.RandomizeSign),
-			Min:           types.Float64PointerValue(config.GenerateFloat64Config.Min),
-			Max:           types.Float64PointerValue(config.GenerateFloat64Config.Max),
-			Precision:     types.Int64PointerValue(config.GenerateFloat64Config.Precision),
-		}
-	case *mgmtv1alpha1.TransformerConfig_GenerateFullAddressConfig:
-		tconfig.GenerateFullAddress = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateFullNameConfig:
-		tconfig.GenerateFullName = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateGenderConfig:
-		tconfig.GenerateGender = &GenerateGender{
-			Abbreviate: types.BoolPointerValue(config.GenerateGenderConfig.Abbreviate),
-		}
-	case *mgmtv1alpha1.TransformerConfig_GenerateInt64PhoneNumberConfig:
-		tconfig.GenerateInt64PhoneNumber = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateInt64Config:
-		tconfig.GenerateInt64 = &GenerateInt64{
-			RandomizeSign: types.BoolPointerValue(config.GenerateInt64Config.RandomizeSign),
-			Min:           types.Int64PointerValue(config.GenerateInt64Config.Min),
-			Max:           types.Int64PointerValue(config.GenerateInt64Config.Max),
-		}
-	case *mgmtv1alpha1.TransformerConfig_GenerateLastNameConfig:
-		tconfig.GenerateLastName = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateSha256HashConfig:
-		tconfig.GenerateSha256Hash = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateSsnConfig:
-		tconfig.GenerateSsn = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateStateConfig:
-		tconfig.GenerateState = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateStreetAddressConfig:
-		tconfig.GenerateStreetAddress = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateStringPhoneNumberConfig:
-		tconfig.GenerateStringPhoneNumber = &GenerateStringPhoneNumber{
-			Min: types.Int64PointerValue(config.GenerateStringPhoneNumberConfig.Min),
-			Max: types.Int64PointerValue(config.GenerateStringPhoneNumberConfig.Max),
-		}
-	case *mgmtv1alpha1.TransformerConfig_GenerateStringConfig:
-		tconfig.GenerateString = &GenerateString{
-			Min: types.Int64PointerValue(config.GenerateStringConfig.Min),
-			Max: types.Int64PointerValue(config.GenerateStringConfig.Max),
-		}
-	case *mgmtv1alpha1.TransformerConfig_GenerateUnixtimestampConfig:
-		tconfig.GenerateUnixtimestamp = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateUsernameConfig:
-		tconfig.GenerateUsername = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateUtctimestampConfig:
-		tconfig.GenerateUtctimestamp = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateUuidConfig:
-		tconfig.GenerateUuid = &GenerateUuid{
-			IncludeHyphens: types.BoolPointerValue(config.GenerateUuidConfig.IncludeHyphens),
-		}
-	case *mgmtv1alpha1.TransformerConfig_GenerateZipcodeConfig:
-		tconfig.GenerateZipcode = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_TransformE164PhoneNumberConfig:
-		tconfig.TransformE164PhoneNumber = &TransformE164PhoneNumber{
-			PreserveLength: types.BoolPointerValue(config.TransformE164PhoneNumberConfig.PreserveLength),
-		}
-	case *mgmtv1alpha1.TransformerConfig_TransformFirstNameConfig:
-		tconfig.TransformFirstName = &TransformFirstName{
-			PreserveLength: types.BoolPointerValue(config.TransformFirstNameConfig.PreserveLength),
-		}
-	case *mgmtv1alpha1.TransformerConfig_TransformFloat64Config:
-		tconfig.TransformFloat64 = &TransformFloat64{
-			RandomizationRangeMin: types.Float64PointerValue(config.TransformFloat64Config.RandomizationRangeMin),
-			RandomizationRangeMax: types.Float64PointerValue(config.TransformFloat64Config.RandomizationRangeMax),
-		}
-	case *mgmtv1alpha1.TransformerConfig_TransformFullNameConfig:
-		tconfig.TransformFullName = &TransformFullName{
-			PreserveLength: types.BoolPointerValue(config.TransformFullNameConfig.PreserveLength),
-		}
-	case *mgmtv1alpha1.TransformerConfig_TransformInt64PhoneNumberConfig:
-		tconfig.TransformInt64PhoneNumber = &TransformInt64PhoneNumber{
-			PreserveLength: types.BoolPointerValue(config.TransformInt64PhoneNumberConfig.PreserveLength),
-		}
-	case *mgmtv1alpha1.TransformerConfig_TransformInt64Config:
-		tconfig.TransformInt64 = &TransformInt64{
-			RandomizationRangeMin: types.Int64PointerValue(config.TransformInt64Config.RandomizationRangeMin),
-			RandomizationRangeMax: types.Int64PointerValue(config.TransformInt64Config.RandomizationRangeMax),
-		}
-	case *mgmtv1alpha1.TransformerConfig_TransformLastNameConfig:
-		tconfig.TransformLastName = &TransformLastName{
-			PreserveLength: types.BoolPointerValue(config.TransformLastNameConfig.PreserveLength),
-		}
-	case *mgmtv1alpha1.TransformerConfig_TransformPhoneNumberConfig:
-		tconfig.TransformPhoneNumber = &TransformPhoneNumber{
-			PreserveLength: types.BoolPointerValue(config.TransformPhoneNumberConfig.PreserveLength),
-		}
-	case *mgmtv1alpha1.TransformerConfig_TransformStringConfig:
-		tconfig.TransformString = &TransformString{
-			PreserveLength: types.BoolPointerValue(config.TransformStringConfig.PreserveLength),
-		}
-	case *mgmtv1alpha1.TransformerConfig_PassthroughConfig:
-		tconfig.Passthrough = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_Nullconfig:
-		tconfig.Null = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_UserDefinedTransformerConfig:
-		tconfig.UserDefinedTransformer = &UserDefinedTransformer{
-			Id: types.StringValue(config.UserDefinedTransformerConfig.Id),
-		}
-	case *mgmtv1alpha1.TransformerConfig_GenerateDefaultConfig:
-		tconfig.GenerateDefault = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_TransformJavascriptConfig:
-		tconfig.TransformJavascript = &TransformJavascript{
-			Code: types.StringValue(config.TransformJavascriptConfig.Code),
-		}
-	case *mgmtv1alpha1.TransformerConfig_GenerateCategoricalConfig:
-		tconfig.GenerateCategorical = &GenerateCategorical{
-			Categories: types.StringPointerValue(config.GenerateCategoricalConfig.Categories),
-		}
-	case *mgmtv1alpha1.TransformerConfig_TransformCharacterScrambleConfig:
-		tconfig.TransformCharacterScramble = &TransformerEmpty{}
-	case *mgmtv1alpha1.TransformerConfig_GenerateJavascriptConfig:
-		tconfig.GenerateJavascript = &GenerateJavascript{
-			Code: types.StringValue(config.GenerateJavascriptConfig.Code),
-		}
-	default:
-		return nil, fmt.Errorf("this job mapping transformer is not currently supported by this provider: %w", errors.ErrUnsupported)
-	}
-	return tconfig, nil
 }
 
 func i32Toi64(input *int32) *int64 {
@@ -973,136 +594,10 @@ type MysqlDestinationTruncateTable struct {
 type JobDestinationAwsS3Options struct{}
 
 type JobMapping struct {
-	Schema      types.String `tfsdk:"schema"`
-	Table       types.String `tfsdk:"table"`
-	Column      types.String `tfsdk:"column"`
-	Transformer *Transformer `tfsdk:"transformer"`
-}
-type Transformer struct {
-	Config *TransformerConfig `tfsdk:"config"`
-}
-
-type TransformerConfig struct {
-	GenerateEmail              *TransformerEmpty          `tfsdk:"generate_email"`
-	TransformEmail             *TransformEmail            `tfsdk:"transform_email"`
-	GenerateBool               *TransformerEmpty          `tfsdk:"generate_bool"`
-	GenerateCardNumber         *GenerateCardNumber        `tfsdk:"generate_card_number"`
-	GenerateCity               *TransformerEmpty          `tfsdk:"generate_city"`
-	GenerateE164PhoneNumber    *GenerateE164PhoneNumber   `tfsdk:"generate_e164_phone_number"`
-	GenerateFirstName          *TransformerEmpty          `tfsdk:"generate_firstname"`
-	GenerateFloat64            *GenerateFloat64           `tfsdk:"generate_float64"`
-	GenerateFullAddress        *TransformerEmpty          `tfsdk:"generate_full_address"`
-	GenerateFullName           *TransformerEmpty          `tfsdk:"generate_fullname"`
-	GenerateGender             *GenerateGender            `tfsdk:"generate_gender"`
-	GenerateInt64PhoneNumber   *TransformerEmpty          `tfsdk:"generate_int64_phone_number"`
-	GenerateInt64              *GenerateInt64             `tfsdk:"generate_int64"`
-	GenerateLastName           *TransformerEmpty          `tfsdk:"generate_lastname"`
-	GenerateSha256Hash         *TransformerEmpty          `tfsdk:"generate_sha256"`
-	GenerateSsn                *TransformerEmpty          `tfsdk:"generate_ssn"`
-	GenerateState              *TransformerEmpty          `tfsdk:"generate_state"`
-	GenerateStreetAddress      *TransformerEmpty          `tfsdk:"generate_street_address"`
-	GenerateStringPhoneNumber  *GenerateStringPhoneNumber `tfsdk:"generate_string_phone_number"`
-	GenerateString             *GenerateString            `tfsdk:"generate_string"`
-	GenerateUnixtimestamp      *TransformerEmpty          `tfsdk:"generate_unix_timestamp"`
-	GenerateUsername           *TransformerEmpty          `tfsdk:"generate_username"`
-	GenerateUtctimestamp       *TransformerEmpty          `tfsdk:"generate_utc_timestamp"`
-	GenerateUuid               *GenerateUuid              `tfsdk:"generate_uuid"`
-	GenerateZipcode            *TransformerEmpty          `tfsdk:"generate_zipcode"`
-	TransformE164PhoneNumber   *TransformE164PhoneNumber  `tfsdk:"transform_e164_phone_number"`
-	TransformFirstName         *TransformFirstName        `tfsdk:"transform_firstname"`
-	TransformFloat64           *TransformFloat64          `tfsdk:"transform_float64"`
-	TransformFullName          *TransformFullName         `tfsdk:"transform_fullname"`
-	TransformInt64PhoneNumber  *TransformInt64PhoneNumber `tfsdk:"transform_int64_phone_number"`
-	TransformInt64             *TransformInt64            `tfsdk:"transform_int64"`
-	TransformLastName          *TransformLastName         `tfsdk:"transform_lastname"`
-	TransformPhoneNumber       *TransformPhoneNumber      `tfsdk:"transform_phone_number"`
-	TransformString            *TransformString           `tfsdk:"transform_string"`
-	Passthrough                *TransformerEmpty          `tfsdk:"passthrough"`
-	Null                       *TransformerEmpty          `tfsdk:"null"`
-	UserDefinedTransformer     *UserDefinedTransformer    `tfsdk:"user_defined_transformer"`
-	GenerateDefault            *TransformerEmpty          `tfsdk:"generate_default"`
-	TransformJavascript        *TransformJavascript       `tfsdk:"transform_javascript"`
-	GenerateJavascript         *GenerateJavascript        `tfsdk:"generate_javascript"`
-	GenerateCategorical        *GenerateCategorical       `tfsdk:"generate_categorical"`
-	TransformCharacterScramble *TransformerEmpty          `tfsdk:"transform_character_scramble"`
-}
-type TransformerEmpty struct{}
-type TransformEmail struct {
-	PreserveDomain types.Bool `tfsdk:"preserve_domain"`
-	PreserveLength types.Bool `tfsdk:"preserve_length"`
-}
-type GenerateCardNumber struct {
-	ValidLuhn types.Bool `tfsdk:"valid_luhn"`
-}
-type GenerateE164PhoneNumber struct {
-	Min types.Int64 `tfsdk:"min"`
-	Max types.Int64 `tfsdk:"max"`
-}
-type GenerateFloat64 struct {
-	RandomizeSign types.Bool    `tfsdk:"randomize_sign"`
-	Min           types.Float64 `tfsdk:"min"`
-	Max           types.Float64 `tfsdk:"max"`
-	Precision     types.Int64   `tfsdk:"precision"`
-}
-type GenerateGender struct {
-	Abbreviate types.Bool `tfsdk:"abbreviate"`
-}
-type GenerateInt64 struct {
-	RandomizeSign types.Bool  `tfsdk:"randomize_sign"`
-	Min           types.Int64 `tfsdk:"min"`
-	Max           types.Int64 `tfsdk:"max"`
-}
-type GenerateStringPhoneNumber struct {
-	Min types.Int64 `tfsdk:"min"`
-	Max types.Int64 `tfsdk:"max"`
-}
-type GenerateString struct {
-	Min types.Int64 `tfsdk:"min"`
-	Max types.Int64 `tfsdk:"max"`
-}
-type GenerateUuid struct {
-	IncludeHyphens types.Bool `tfsdk:"include_hyphens"`
-}
-type TransformE164PhoneNumber struct {
-	PreserveLength types.Bool `tfsdk:"preserve_length"`
-}
-type TransformFirstName struct {
-	PreserveLength types.Bool `tfsdk:"preserve_length"`
-}
-type TransformFloat64 struct {
-	RandomizationRangeMin types.Float64 `tfsdk:"randomization_range_min"`
-	RandomizationRangeMax types.Float64 `tfsdk:"randomization_range_max"`
-}
-type TransformFullName struct {
-	PreserveLength types.Bool `tfsdk:"preserve_length"`
-}
-type TransformInt64PhoneNumber struct {
-	PreserveLength types.Bool `tfsdk:"preserve_length"`
-}
-type TransformInt64 struct {
-	RandomizationRangeMin types.Int64 `tfsdk:"randomization_range_min"`
-	RandomizationRangeMax types.Int64 `tfsdk:"randomization_range_max"`
-}
-type TransformLastName struct {
-	PreserveLength types.Bool `tfsdk:"preserve_length"`
-}
-type TransformPhoneNumber struct {
-	PreserveLength types.Bool `tfsdk:"preserve_length"`
-}
-type TransformString struct {
-	PreserveLength types.Bool `tfsdk:"preserve_length"`
-}
-type TransformJavascript struct {
-	Code types.String `tfsdk:"code"`
-}
-type GenerateJavascript struct {
-	Code types.String `tfsdk:"code"`
-}
-type UserDefinedTransformer struct {
-	Id types.String `tfsdk:"id"`
-}
-type GenerateCategorical struct {
-	Categories types.String `tfsdk:"categories"`
+	Schema      types.String                   `tfsdk:"schema"`
+	Table       types.String                   `tfsdk:"table"`
+	Column      types.String                   `tfsdk:"column"`
+	Transformer *transformer_model.Transformer `tfsdk:"transformer"`
 }
 
 type ActivityOptions struct {
